Refactor: Split monolithic main file into logical modules
Fixes #5

Changes:
- Split e3dcset.cpp (2653 lines) into 5 logical modules:
  * config.cpp/h - Configuration, argument parsing, connection setup
  * rscp_handler.cpp/h - RSCP protocol communication, request/response handling
  * output.cpp/h - Output functions (text & JSON), tag handling
  * history.cpp/h - History data processing, timestamp handling
  * e3dcset.cpp - Main entry point, argument parsing

- Updated Makefile to compile all modules
- All existing tests pass (33/33)
- No functional changes - pure code organization
- Maintains 1:1 compatibility with original implementation

Benefits:
- Improved maintainability
- Better code organization
- Easier to navigate and understand
- Foundation for future development
